+++ rvc/train/extract/extract.py
@@ -14,7 +14,10 @@ from distutils.util import strtobool
 
 now_dir = os.getcwd()
 sys.path.append(os.path.join(now_dir))
-import rvc.train.zluda
+
+# Zluda hijack
+import rvc.lib.zluda
+
 from rvc.lib.utils import load_audio, load_embedding
 from rvc.train.extract.preparing_files import generate_config, generate_filelist
 from rvc.lib.predictors.RMVPE import RMVPE0Predictor
